+++ src/display/e_paper_154.h
@@ -29,7 +29,7 @@ extern int curHour, curMinute, curDay, curMonth, curYear;
 
 void display_init();
 void display_first_time(String input);
-void display_update(String input, String msg, int packet_rate, int packets_count, unsigned long total_attack_counter);
+void display_update(String input, String msg, char spin, int packet_rate, int packets_count, unsigned long total_attack_counter);
 #endif
 
 #endif