+++ ChangeLog.md
@@ -3,6 +3,7 @@ ADD:
  * Multiplatform in .ini 
  * Start adding support for MiniTV 1.44" 
  * MiniTV show time, date, status.
+ * Fix total packets counter
 
 -=# v0.0.4 (20260208) #=- 
 ADD: